+++ src/main/java/org/bdyb/hotel/init/InitCustomers.java
@@ -27,7 +27,7 @@ public class InitCustomers {
                 customerService.addOne(CustomerDto.builder()
                         .firstName(name)
                         .lastName(surname)
-                        .pesel(i)
+                        .pesel((long) i)
                         .birthday(birthday.plusYears(i).toDate())
                         .identityCard(IdentityCardDto.builder()
                                 .type(IdCardType.ID_CARD)
